+++ EldredBrown.ProFootball.NETCore.Services/Calculator.cs
@@ -35,7 +35,7 @@ namespace EldredBrown.ProFootball.NETCore.Services
         /// <param name="x">The first operand.</param>
         /// <param name="y">The second operand.</param>
         /// <returns>The result of x * y.</returns>
-		public virtual decimal Multiply(decimal x, decimal y)
+		public virtual double Multiply(double x, double y)
         {
             return x * y;
         }
@@ -46,11 +46,11 @@ namespace EldredBrown.ProFootball.NETCore.Services
         /// <param name="x">The numerator.</param>
         /// <param name="y">The denominator.</param>
         /// <returns>The result of x / y; null if the denominator is zero.</returns>
-        public virtual decimal? Divide(decimal x, decimal y)
+        public virtual double? Divide(double x, double y)
         {
             // Rather than throw an error for division by zero, 
             // this will return a default result of null if division by zero occurs.
-            decimal? result = null;
+            double? result = null;
 
             if (y != 0)
             {
@@ -65,7 +65,7 @@ namespace EldredBrown.ProFootball.NETCore.Services
         /// </summary>
         /// <param name="teamSeason">The <see cref="TeamSeason"/> object for which a winning percentage will be calculated.</param>
         /// <returns>The <see cref="TeamSeason"/> object's winning percentage.</returns>
-        public virtual decimal? CalculateWinningPercentage(TeamSeason teamSeason)
+        public virtual double? CalculateWinningPercentage(TeamSeason teamSeason)
         {
             var result = Divide(2 * teamSeason.Wins + teamSeason.Ties, 2 * teamSeason.Games);
 
@@ -77,11 +77,10 @@ namespace EldredBrown.ProFootball.NETCore.Services
         /// </summary>
         /// <param name="teamSeason">The <see cref="TeamSeason"/> object for which a Pythagorean winning percentage will be calculated.</param>
         /// <returns>The <see cref="TeamSeason"/> object's Pythagorean winning percentage.</returns>
-        public virtual decimal? CalculatePythagoreanWinningPercentage(TeamSeason teamSeason)
+        public virtual double? CalculatePythagoreanWinningPercentage(TeamSeason teamSeason)
         {
-            var x = (decimal)Math.Pow(teamSeason.PointsFor, _exponent);
-            var y = 
-                (decimal)(Math.Pow(teamSeason.PointsFor, _exponent) + Math.Pow(teamSeason.PointsAgainst, _exponent));
+            var x = Math.Pow(teamSeason.PointsFor, _exponent);
+            var y = (Math.Pow(teamSeason.PointsFor, _exponent) + Math.Pow(teamSeason.PointsAgainst, _exponent));
             var pct = Divide(x, y);
 
             return pct;
